DBの項目を変更

+++ database/migrations/2023_06_28_144920_create_event_participants_table.php
@@ -16,7 +16,7 @@ class CreateEventParticipantsTable extends Migration
         Schema::create('event_participants', function (Blueprint $table) {
             $table->id();
             $table->unsignedBigInteger('user_id');
-            $table->unsignedInteger('event_id');
+            $table->unsignedBigInteger('event_id');
             $table->timestamps();
 
             $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
